chore(wfs): move WFS related code into source folder (refs #46)
